+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version 0.5.1
 
 * Fixed some Python 3 incompatibilities
 * Fixed crash if DBus library is not installed
+* Fixed support for newer version of Banshee
 
 Version 0.5
 -----------